+++ Editor/UI/Framework/Draw2D.cs
@@ -4,7 +4,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
 {
     internal class Draw2D
     {
-        internal enum Origin
+        public enum Origin
         {
             TopLeft,
             BottomLeft
@@ -19,7 +19,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
         Vector4 m_ClipRect;
         bool m_ClipRectEnabled = false;
 
-        internal Draw2D(string shaderName)
+        public Draw2D(string shaderName)
         {
             m_ShaderName = shaderName;
             CheckAndSetupMaterial();
@@ -52,7 +52,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             return m_Material != null;
         }
 
-        internal void OnGUI()
+        public void OnGUI()
         {
             if (m_GLStyle == null)
             {
@@ -62,35 +62,35 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             }
         }
 
-        internal void SetClipRect(Rect clipRect)
+        public void SetClipRect(Rect clipRect)
         {
             m_ClipRect = new Vector4(clipRect.x, clipRect.y, clipRect.x + clipRect.width, clipRect.y + clipRect.height);
             m_ClipRectEnabled = true;
 
             if (CheckAndSetupMaterial())
-                return;
-
-            m_Material.SetFloat("_UseClipRect", m_ClipRectEnabled ? 1f : 0f);
-            m_Material.SetVector("_ClipRect", m_ClipRect);
+            {
+                m_Material.SetFloat("_UseClipRect", m_ClipRectEnabled ? 1f : 0f);
+                m_Material.SetVector("_ClipRect", m_ClipRect);
+            }
         }
 
-        internal void ClearClipRect()
+        public void ClearClipRect()
         {
             m_ClipRectEnabled = false;
 
             if (CheckAndSetupMaterial())
-                return;
-
-            m_Material.SetFloat("_UseClipRect", m_ClipRectEnabled ? 1f : 0f);
-            m_Material.SetVector("_ClipRect", m_ClipRect);
+            {
+                m_Material.SetFloat("_UseClipRect", m_ClipRectEnabled ? 1f : 0f);
+                m_Material.SetVector("_ClipRect", m_ClipRect);
+            }
         }
 
-        internal Rect GetClipRect()
+        public Rect GetClipRect()
         {
             return new Rect(m_ClipRect.x, m_ClipRect.y, m_ClipRect.z - m_ClipRect.x, m_ClipRect.w - m_ClipRect.y);
         }
 
-        internal bool DrawStart(Rect r, Origin origin = Origin.TopLeft)
+        public bool DrawStart(Rect r, Origin origin = Origin.TopLeft)
         {
             if (Event.current.type != EventType.Repaint)
                 return false;
@@ -105,13 +105,13 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             return true;
         }
 
-        internal bool DrawStart(float w, float h, Origin origin = Origin.TopLeft, GUIStyle style = null)
+        public bool DrawStart(float w, float h, Origin origin = Origin.TopLeft, GUIStyle style = null)
         {
             Rect r = GUILayoutUtility.GetRect(w, h, style == null ? m_GLStyle : style, GUILayout.ExpandWidth(false), GUILayout.ExpandHeight(false));
             return DrawStart(r, origin);
         }
 
-        internal void DrawEnd()
+        public void DrawEnd()
         {
         }
 
@@ -130,7 +130,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             }
         }
 
-        internal void DrawFilledBox(float x, float y, float w, float h, Color col)
+        public void DrawFilledBox(float x, float y, float w, float h, Color col)
         {
             float x2 = x + w;
             float y2 = y + h;
@@ -160,12 +160,61 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             }
         }
 
-        internal void DrawFilledBox(float x, float y, float w, float h, float r, float g, float b)
+        public void DrawFilledBox(float x, float y, float w, float h, float r, float g, float b)
         {
             DrawFilledBox(x, y, w, h, new Color(r, g, b));
         }
 
-        internal void DrawLine(float x, float y, float x2, float y2, Color col)
+        public void DrawFilledCircle(float x, float y, float r, Color col, float quality = 12)
+        {
+            Translate(ref x, ref y);
+
+            GL.Begin(GL.TRIANGLES);
+            GL.Color(col);
+
+            float x1 = x + r * Mathf.Cos(0);
+            float y1 = y + r * Mathf.Sin(0);
+            float step = Mathf.PI / quality;
+            for (float angle = 0; angle < Mathf.PI * 2; angle += step)
+            {
+                float nextAngle = angle + step;
+                float x2 = x + r * Mathf.Cos(nextAngle);
+                float y2 = y + r * Mathf.Sin(nextAngle);
+                GL.Vertex3(x, y, 0);
+                GL.Vertex3(x1, y1, 0);
+                GL.Vertex3(x2, y2, 0);
+                x1 = x2;
+                y1 = y2;
+            }
+
+            GL.End();
+        }
+
+        public void DrawCircle(float x, float y, float r, Color col, float quality = 12)
+        {
+            Translate(ref x, ref y);
+
+            GL.Begin(GL.LINES);
+            GL.Color(col);
+
+            float x1 = x + r * Mathf.Cos(0);
+            float y1 = y + r * Mathf.Sin(0);
+            float step = Mathf.PI / quality;
+            for (float angle = 0; angle < Mathf.PI * 2; angle += step)
+            {
+                float nextAngle = angle + step;
+                float x2 = x + r * Mathf.Cos(nextAngle);
+                float y2 = y + r * Mathf.Sin(nextAngle);
+                GL.Vertex3(x1, y1, 0);
+                GL.Vertex3(x2, y2, 0);
+                x1 = x2;
+                y1 = y2;
+            }
+
+            GL.End();
+        }
+
+        public void DrawLine(float x, float y, float x2, float y2, Color col)
         {
             Translate(ref x, ref y);
             Translate(ref x2, ref y2);
@@ -177,12 +226,12 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             GL.End();
         }
 
-        internal void DrawLine(float x, float y, float x2, float y2, float r, float g, float b)
+        public void DrawLine(float x, float y, float x2, float y2, float r, float g, float b)
         {
             DrawLine(x, y, x2, y2, new Color(r, g, b));
         }
 
-        internal void DrawBox(float x, float y, float w, float h, Color col)
+        public void DrawBox(float x, float y, float w, float h, Color col)
         {
             float x2 = x + w;
             float y2 = y + h;
@@ -200,7 +249,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             GL.End();
         }
 
-        internal void DrawBox(float x, float y, float w, float h, float r, float g, float b)
+        public void DrawBox(float x, float y, float w, float h, float r, float g, float b)
         {
             DrawBox(x, y, w, h, new Color(r, g, b));
         }